# api/openapi.yaml
openapi: "3.1.0"
info:
  title: DirtOS REST API
  version: "1"
  description: |
    Local REST API embedded in the DirtOS desktop application.
    Intended for plugin development and 3rd-party integrations (n8n, Home Assistant, scripts, etc.).

    **Base URL:** `http://127.0.0.1:7272`
    Override the port with the `DIRTOS_API_PORT` environment variable before launching DirtOS.

    All endpoints return JSON. Timestamps are ISO 8601 strings in UTC.
    Enum values are lowercase snake_case strings unless noted otherwise.
servers:
  - url: http://127.0.0.1:7272
    description: DirtOS local API (default port)
...
paths:
...
  /api/v1/environments:
    get:
      tags: [Environments]
      summary: List environments
      operationId: listEnvironments
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of environments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Environment"
    post:
      tags: [Environments]
      summary: Create an environment
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEnvironment"
      responses:
        "200":
          description: Created environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
  /api/v1/environments/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Environments]
      summary: Get an environment
      operationId: getEnvironment
      responses:
        "200":
          description: Environment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Environments]
      summary: Update an environment
      operationId: updateEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvironment"
      responses:
        "200":
          description: Updated environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Environment"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Environments]
      summary: Delete an environment
      operationId: deleteEnvironment
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Locations
  # ---------------------------------------------------------------------------
...
  /api/v1/issues:
    get:
      tags: [Issues]
      summary: List issues in an environment
      operationId: listIssues
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      tags: [Issues]
      summary: Create an issue
      operationId: createIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewIssue"
      responses:
        "200":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /api/v1/issues/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Issues]
      summary: Get an issue
      operationId: getIssue
      responses:
        "200":
          description: Issue record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Issues]
      summary: Update an issue
      operationId: updateIssue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssue"
      responses:
        "200":
          description: Updated issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Issues]
      summary: Delete an issue
      operationId: deleteIssue
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Journal
  # ---------------------------------------------------------------------------
  /api/v1/journal:
    get:
      tags: [Journal]
      summary: List journal entries in an environment
      operationId: listJournalEntries
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      tags: [Journal]
      summary: Create a journal entry
      operationId: createJournalEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewJournalEntry"
      responses:
        "200":
          description: Created journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /api/v1/journal/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Journal]
      summary: Get a journal entry
      operationId: getJournalEntry
      responses:
        "200":
          description: Journal entry record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Journal]
      summary: Update a journal entry
      operationId: updateJournalEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJournalEntry"
      responses:
        "200":
          description: Updated journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Journal]
      summary: Delete a journal entry
      operationId: deleteJournalEntry
      responses:
        "204":
          description: Deleted successfully

  # ---------------------------------------------------------------------------
  # Harvests
  # ---------------------------------------------------------------------------
  /api/v1/harvests:
    get:
      tags: [Harvests]
      summary: List harvests in an environment
      operationId: listHarvests
      parameters:
        - $ref: "#/components/parameters/environment_id"
        - name: date_from
          in: query
          schema:
            type: string
            format: date
            example: "2026-01-01"
          description: Filter harvests on or after this date (YYYY-MM-DD)
        - name: date_to
          in: query
          schema:
            type: string
            format: date
            example: "2026-12-31"
          description: Filter harvests on or before this date (YYYY-MM-DD)
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: List of harvests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Harvest"
    post:
      tags: [Harvests]
      summary: Record a harvest
      operationId: createHarvest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewHarvest"
      responses:
        "200":
          description: Created harvest record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Harvest"
  /api/v1/harvests/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags: [Harvests]
      summary: Get a harvest record
      operationId: getHarvest
      responses:
        "200":
          description: Harvest record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Harvest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Harvests]
      summary: Delete a harvest record
      operationId: deleteHarvest
      responses:
        "204":
          description: Deleted successfully

# ===========================================================================
# Components
# ===========================================================================
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    environment_id:
      name: environment_id
      in: query
      required: true
      schema:
        type: integer
        format: int64
      description: Filter results to a specific environment
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
      description: Maximum number of records to return
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of records to skip (for pagination)
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: not found
  schemas:
...
    Environment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        elevation_m:
          type: number
          nullable: true
        timezone:
          type: string
          nullable: true
        climate_zone:
          type: string
          nullable: true
        asset_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewEnvironment:
      type: object
      required: [name]
      properties:
        name:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        elevation_m:
          type: number
          nullable: true
        timezone:
          type: string
          nullable: true
        climate_zone:
          type: string
          nullable: true
    UpdateEnvironment:
      type: object
      properties:
        name:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        elevation_m:
          type: number
          nullable: true
        timezone:
          type: string
          nullable: true
        climate_zone:
          type: string
          nullable: true

    # -------------------------------------------------------------------------
    # Location
    # -------------------------------------------------------------------------
...
    IssueStatus:
      type: string
      enum: [new, open, in_progress, closed]
    IssuePriority:
      type: string
      enum: [low, medium, high, critical]
    Issue:
      type: object
      properties:
        id:
          type: integer
          format: int64
        environment_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/IssueStatus"
        priority:
          $ref: "#/components/schemas/IssuePriority"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
    NewIssue:
      type: object
      required: [title]
      properties:
        environment_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/IssueStatus"
        priority:
          $ref: "#/components/schemas/IssuePriority"
    UpdateIssue:
      type: object
      properties:
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/IssueStatus"
        priority:
          $ref: "#/components/schemas/IssuePriority"
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true

    # -------------------------------------------------------------------------
    # Journal
    # -------------------------------------------------------------------------
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        environment_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        body:
          type: string
          nullable: true
        conditions_json:
          type: string
          nullable: true
          description: JSON blob with environmental conditions at time of entry
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    NewJournalEntry:
      type: object
      required: [title]
      properties:
        environment_id:
          type: integer
          format: int64
          nullable: true
        plant_id:
          type: integer
          format: int64
          nullable: true
        location_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        body:
          type: string
          nullable: true
        conditions_json:
          type: string
          nullable: true
    UpdateJournalEntry:
      type: object
      properties:
        title:
          type: string
          nullable: true
        body:
          type: string
          nullable: true
        conditions_json:
          type: string
          nullable: true

    # -------------------------------------------------------------------------
    # Harvest
    # -------------------------------------------------------------------------
    Harvest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        plant_id:
          type: integer
          format: int64
        harvest_date:
          type: string
          format: date
        quantity:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
        quality_rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        notes:
          type: string
          nullable: true
        asset_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    NewHarvest:
      type: object
      required: [plant_id, harvest_date]
      properties:
        plant_id:
          type: integer
          format: int64
        harvest_date:
          type: string
          format: date
          example: "2026-07-15"
        quantity:
          type: number
          nullable: true
        unit:
          type: string
          nullable: true
          example: kg
        quality_rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        notes:
          type: string
          nullable: true